Remove beta prefix
