--- Sources/ProcessOut/Sources/UI/Modules/NativeAlternativePaymentMethod/BetaViewModel/BetaNativeAlternativePaymentMethodViewModel.swift
@@ -1,12 +1,0 @@
-//
-//  BetaNativeAlternativePaymentMethodViewModel.swift
-//  ProcessOut
-//
-//  Created by Andrii Vysotskyi on 20.04.2023.
-//
-
-protocol BetaNativeAlternativePaymentMethodViewModel: ViewModel<BetaNativeAlternativePaymentMethodViewModelState> {
-
-    /// Submits parameter values.
-    func submit()
-}
