+++ Sources/ProcessOut/Sources/UI/Modules/NativeAlternativePaymentMethod/View/Cells/NativeAlternativePaymentMethodSectionHeaderView.swift
@@ -1,17 +1,17 @@
 //
-//  NativeAlternativePaymentMethodTitleCell.swift
+//  NativeAlternativePaymentMethodSectionHeaderView.swift
 //  ProcessOut
 //
-//  Created by Andrii Vysotskyi on 20.04.2023.
+//  Created by Andrii Vysotskyi on 26.04.2023.
 //
 
 import UIKit
 
-final class NativeAlternativePaymentMethodTitleCell: UICollectionViewCell {
+final class NativeAlternativePaymentMethodSectionHeaderView: UICollectionReusableView {
 
     override init(frame: CGRect) {
         super.init(frame: frame)
-        commonInit()
+        self.commonInit()
     }
 
     @available(*, unavailable)
@@ -19,22 +19,21 @@ final class NativeAlternativePaymentMethodTitleCell: UICollectionViewCell {
         fatalError("init(coder:) has not been implemented")
     }
 
-    func configure(item: BetaNativeAlternativePaymentMethodViewModelState.TitleItem, style: POTextStyle?) {
+    func configure(item: NativeAlternativePaymentMethodViewModelState.SectionIdentifier, style: POTextStyle?) {
         let style = style ?? Constants.defaultStyle
         titleLabel.attributedText = AttributedStringBuilder()
             .typography(style.typography)
-            .textStyle(textStyle: .title1)
-            .alignment(.center)
-            .lineBreakMode(.byWordWrapping)
+            .textStyle(textStyle: .body)
             .textColor(style.color)
-            .string(item.text)
+            .alignment(.center)
+            .string(item.title ?? "")
             .build()
     }
 
     // MARK: - Private Nested Types
 
     private enum Constants {
-        static let defaultStyle = POTextStyle(color: Asset.Colors.Text.primary.color, typography: .title)
+        static let defaultStyle = POTextStyle(color: Asset.Colors.Text.primary.color, typography: .bodyLarge)
     }
 
     // MARK: - Private Properties
@@ -44,20 +43,20 @@ final class NativeAlternativePaymentMethodTitleCell: UICollectionViewCell {
         label.translatesAutoresizingMaskIntoConstraints = false
         label.adjustsFontForContentSizeCategory = false
         label.numberOfLines = 0
-        label.setContentCompressionResistancePriority(.required, for: .vertical)
         label.setContentHuggingPriority(.required, for: .vertical)
+        label.setContentCompressionResistancePriority(.required, for: .vertical)
         return label
     }()
 
     // MARK: - Private Methods
 
     private func commonInit() {
-        contentView.addSubview(titleLabel)
+        addSubview(titleLabel)
         let constraints = [
-            titleLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
-            titleLabel.centerXAnchor.constraint(equalTo: contentView.centerXAnchor),
-            titleLabel.centerYAnchor.constraint(equalTo: contentView.centerYAnchor),
-            titleLabel.topAnchor.constraint(equalTo: contentView.topAnchor)
+            titleLabel.leadingAnchor.constraint(equalTo: leadingAnchor),
+            titleLabel.centerXAnchor.constraint(equalTo: centerXAnchor),
+            titleLabel.topAnchor.constraint(equalTo: topAnchor),
+            titleLabel.centerYAnchor.constraint(equalTo: centerYAnchor)
         ]
         NSLayoutConstraint.activate(constraints)
     }
