+++ Sources/ProcessOut/Sources/UI/Modules/NativeAlternativePaymentMethod/View/Cells/Submitted/NativeAlternativePaymentMethodSubmittedCellStyle.swift
