+++ Sources/ProcessOut/Sources/UI/Modules/NativeAlternativePaymentMethod/ViewModel/DefaultNativeAlternativePaymentMethodViewModel.swift
@@ -2,11 +2,12 @@
 //  DefaultNativeAlternativePaymentMethodViewModel.swift
 //  ProcessOut
 //
-//  Created by Andrii Vysotskyi on 19.10.2022.
+//  Created by Andrii Vysotskyi on 20.04.2023.
 //
 
 import Foundation
 
+// swiftlint:disable:next type_body_length
 final class DefaultNativeAlternativePaymentMethodViewModel:
     BaseViewModel<NativeAlternativePaymentMethodViewModelState>, NativeAlternativePaymentMethodViewModel {
 
@@ -18,6 +19,8 @@ final class DefaultNativeAlternativePaymentMethodViewModel:
         self.interactor = interactor
         self.configuration = configuration
         self.completion = completion
+        inputValuesObservations = []
+        inputValuesCache = [:]
         super.init(state: .idle)
         observeInteractorStateChanges()
     }
@@ -37,6 +40,18 @@ final class DefaultNativeAlternativePaymentMethodViewModel:
 
     private enum Constants {
         static let captureSuccessCompletionDelay: TimeInterval = 3
+        static let maximumCodeLength = 6
+    }
+
+    private struct InputValue {
+
+        /// Indicates whether parameter is invalid.
+        @ReferenceWrapper
+        var isInvalid: Bool
+
+        /// Current parameter's value. This value won't be modified by view model.
+        @ReferenceWrapper
+        var value: String
     }
 
     // MARK: - NativeAlternativePaymentMethodInteractor
@@ -53,6 +68,9 @@ final class DefaultNativeAlternativePaymentMethodViewModel:
         return formatter
     }()
 
+    private var inputValuesCache: [String: InputValue]
+    private var inputValuesObservations: [AnyObject]
+
     // MARK: - Private Methods
 
     private func observeInteractorStateChanges() {
@@ -64,13 +82,17 @@ final class DefaultNativeAlternativePaymentMethodViewModel:
         case .idle:
             state = .idle
         case .starting:
-            state = .loading
+            let sections = [
+                State.Section(id: .init(id: nil, title: nil), items: [.loader])
+            ]
+            let startedState = State.Started(sections: sections, actions: nil, isEditingAllowed: false)
+            state = .started(startedState)
         case .started(let startedState):
-            state = convertToState(startedState: startedState)
+            state = convertToState(startedState: startedState, isSubmitting: false)
         case .failure(let failure):
             completion?(.failure(failure))
-        case .submitting(let startedStateSnapshot):
-            state = convertToState(startedState: startedStateSnapshot, isSubmitting: true)
+        case .submitting(let startedState):
+            state = convertToState(startedState: startedState, isSubmitting: true)
         case .submitted:
             completion?(.success(()))
         case .awaitingCapture(let awaitingCaptureState):
@@ -80,62 +102,63 @@ final class DefaultNativeAlternativePaymentMethodViewModel:
         }
     }
 
-    private func convertToState(
-        startedState: InteractorState.Started, isSubmitting: Bool = false
-    ) -> State {
-        let parameters = startedState.parameters.map { parameter -> State.Parameter in
-            let value = startedState.values[parameter.key]
-            let parameterValue: String
-            if case .singleSelect = parameter.type {
-                // Value of single select parameter is not user friendly instead display name should be used
-                parameterValue = parameter.availableValues?.first { $0.value == value?.value }?.displayName ?? ""
-            } else {
-                parameterValue = value?.value ?? ""
-            }
-            let viewModel = State.Parameter(
-                name: parameter.displayName,
-                placeholder: placeholder(for: parameter),
-                value: parameterValue,
-                type: parameter.type,
-                length: parameter.length,
-                availableValues: parameter.availableValues?.map { availableValue in
-                    State.AvailableParameterValue(
-                        name: availableValue.displayName,
-                        value: availableValue.value,
-                        isSelected: availableValue.value == value?.value
-                    )
-                } ?? [],
-                errorMessage: value?.recentErrorMessage,
-                update: { [weak self] newValue in
-                    _ = self?.interactor.updateValue(newValue, for: parameter.key)
-                },
-                formatted: { [weak self] value in
-                    self?.interactor.formatted(value: value, type: parameter.type) ?? ""
-                }
-            )
-            return viewModel
-        }
-        let state = State.Started(
-            title: configuration.title ?? Strings.title(startedState.gatewayDisplayName),
-            parameters: parameters,
-            isSubmitting: isSubmitting,
-            primaryAction: submitAction(startedState: startedState, isSubmitting: isSubmitting),
-            secondaryAction: cancelAction(isEnabled: !isSubmitting)
+    private func convertToState(startedState: InteractorState.Started, isSubmitting: Bool) -> State {
+        let titleItem = State.TitleItem(
+            text: configuration.title ?? Strings.title(startedState.gatewayDisplayName)
         )
-        return .started(state)
+        var sections = [
+            State.Section(id: .init(id: nil, title: nil), items: [.title(titleItem)])
+        ]
+        for (offset, parameter) in startedState.parameters.enumerated() {
+            let value = startedState.values[parameter.key] ?? .init(value: nil, recentErrorMessage: nil)
+            var items = [
+                createItem(
+                    parameter: parameter,
+                    value: value,
+                    isEditingAllowed: !isSubmitting,
+                    isLast: offset == startedState.parameters.indices.last
+                )
+            ]
+            if let message = value.recentErrorMessage {
+                items.append(.error(State.ErrorItem(description: message)))
+            }
+            let section = State.Section(
+                id: .init(id: parameter.key, title: parameter.displayName), items: items
+            )
+            sections.append(section)
+        }
+        let startedState = State.Started(
+            sections: sections,
+            actions: .init(
+                primary: submitAction(startedState: startedState, isSubmitting: isSubmitting),
+                secondary: cancelAction(isEnabled: !isSubmitting)
+            ),
+            isEditingAllowed: !isSubmitting
+        )
+        return .started(startedState)
     }
 
     private func convertToState(awaitingCaptureState: InteractorState.AwaitingCapture) -> State {
-        guard let expectedActionMessage = awaitingCaptureState.expectedActionMessage else {
-            return .loading
+        let item: State.Item
+        if let expectedActionMessage = awaitingCaptureState.expectedActionMessage {
+            let submittedItem = State.SubmittedItem(
+                message: expectedActionMessage,
+                logoImage: awaitingCaptureState.gatewayLogoImage,
+                image: awaitingCaptureState.actionImage,
+                isCaptured: false
+            )
+            item = .submitted(submittedItem)
+        } else {
+            item = .loader
         }
-        let submittedState = State.Submitted(
-            message: expectedActionMessage,
-            logoImage: awaitingCaptureState.gatewayLogoImage,
-            image: awaitingCaptureState.actionImage,
-            isCaptured: false
+        let startedState = State.Started(
+            sections: [
+                .init(id: .init(id: nil, title: nil), items: [item])
+            ],
+            actions: nil,
+            isEditingAllowed: false
         )
-        return .submitted(submittedState)
+        return .started(startedState)
     }
 
     private func configure(with capturedState: InteractorState.Captured) {
@@ -149,28 +172,24 @@ final class DefaultNativeAlternativePaymentMethodViewModel:
                     self?.completion?(.success(()))
                 }
             )
-            let submittedState = State.Submitted(
+            let submittedItem = State.SubmittedItem(
                 message: Strings.Success.message,
                 logoImage: capturedState.gatewayLogo,
                 image: Asset.Images.success.image,
                 isCaptured: true
             )
-            state = .submitted(submittedState)
+            let startedState = State.Started(
+                sections: [
+                    .init(id: .init(id: nil, title: nil), items: [.submitted(submittedItem)])
+                ],
+                actions: nil,
+                isEditingAllowed: false
+            )
+            state = .started(startedState)
         }
     }
 
-    // MARK: - Utils
-
-    private func placeholder(for parameter: PONativeAlternativePaymentMethodParameter) -> String? {
-        switch parameter.type {
-        case .numeric, .text, .singleSelect:
-            return nil
-        case .email:
-            return Strings.Email.placeholder
-        case .phone:
-            return Strings.Phone.placeholder
-        }
-    }
+    // MARK: - Actions
 
     private func submitAction(startedState: InteractorState.Started, isSubmitting: Bool) -> State.Action {
         let title: String
@@ -209,5 +228,85 @@ final class DefaultNativeAlternativePaymentMethodViewModel:
             }
         )
         return action
+    }
+
+    // MARK: - Input Items
+
+    private func createItem(
+        parameter: PONativeAlternativePaymentMethodParameter,
+        value parameterValue: InteractorState.ParameterValue,
+        isEditingAllowed: Bool,
+        isLast: Bool
+    ) -> State.Item {
+        let inputValue: InputValue
+        if let value = inputValuesCache[parameter.key] {
+            inputValue = value
+            inputValue.value = parameterValue.value ?? ""
+            inputValue.isInvalid = parameterValue.recentErrorMessage != nil
+        } else {
+            inputValue = InputValue(
+                isInvalid: .init(value: parameterValue.recentErrorMessage != nil),
+                value: .init(value: parameterValue.value ?? "")
+            )
+            inputValuesCache[parameter.key] = inputValue
+            let observer = inputValue.$value.addObserver { [weak self] updatedValue in
+                self?.interactor.updateValue(updatedValue, for: parameter.key)
+            }
+            inputValuesObservations.append(observer)
+        }
+        switch parameter.type {
+        case .numeric where (parameter.length ?? .max) <= Constants.maximumCodeLength:
+            let inputItem = State.CodeInputItem(
+                length: parameter.length!, // swiftlint:disable:this force_unwrapping
+                isInvalid: inputValue.$isInvalid,
+                value: inputValue.$value,
+                isEditingAllowed: isEditingAllowed
+            )
+            return .codeInput(inputItem)
+        case .singleSelect:
+            return createPickerItem(parameter: parameter, value: inputValue)
+        default:
+            let inputItem = State.InputItem(
+                type: parameter.type,
+                placeholder: placeholder(for: parameter),
+                isInvalid: inputValue.$isInvalid,
+                value: inputValue.$value,
+                isEditingAllowed: isEditingAllowed,
+                isLast: isLast,
+                formatted: { [weak self] value in
+                    self?.interactor.formatted(value: value, type: parameter.type) ?? ""
+                }
+            )
+            return .input(inputItem)
+        }
+    }
+
+    private func createPickerItem(
+        parameter: PONativeAlternativePaymentMethodParameter, value: InputValue
+    ) -> State.Item {
+        assert(parameter.type == .singleSelect)
+        let options = parameter.availableValues?.map { option in
+            State.PickerOption(name: option.displayName, isSelected: option.value == value.value) { [weak self] in
+                self?.interactor.updateValue(option.value, for: parameter.key)
+            }
+        }
+        let item = State.PickerItem(
+            // Value of single select parameter is not user friendly instead display name should be used.
+            value: parameter.availableValues?.first { $0.value == value.value }?.displayName ?? "",
+            isInvalid: value.isInvalid,
+            options: options ?? []
+        )
+        return .picker(item)
+    }
+
+    private func placeholder(for parameter: PONativeAlternativePaymentMethodParameter) -> String? {
+        switch parameter.type {
+        case .numeric, .text, .singleSelect:
+            return nil
+        case .email:
+            return Strings.Email.placeholder
+        case .phone:
+            return Strings.Phone.placeholder
+        }
     }
 }
